Adicionar verificação para impedir a exclusão de empresas com veículos cadastrados, adicionar log no servidor e melhorar o tratamento de formulários de login e listagens de usuários e veículos

+++ src/app/api/companies/route.ts
@@ -31,6 +31,17 @@ export async function DELETE(req: Request) {
     );
   }
 
+  const vehicle = await prisma.vehicle.findFirst({
+    where: { companyId: id },
+  });
+
+  if (vehicle) {
+    return NextResponse.json(
+      { error: "Empresa com veículos cadastrados" },
+      { status: 401 }
+    );
+  }
+
   await prisma.company.delete({
     where: { id },
   });
